skills/technical-report-docx/scripts/requirements.txt
pypdf>=5.0.0
python-docx>=1.1.2
